backend/requirements.txt
requests==2.32.2
urllib3==1.26.6
python-dotenv==1.0.1
weaviate-client==4.10.4
flask==2.3.3
ollama>=0.4.5,<0.5.0
numpy==1.26.0
pandas==1.5.3
